feat: Update the project

+++ Projects/product-cost-calculator/client/components/projects/jsx/ResourceTables.jsx
@@ -1,14 +1,19 @@
 "use client";
 
-import { Table } from "@/components/ui";
+import { Modal, Table } from "@/components/ui";
 import styles from "../css/Project.module.css";
 import {
   filterByType,
   RESOURCE_COLUMNS_MAP,
   RESOURCE_ORDER,
 } from "../helper/resourceColumns";
+import { BASE_PATH } from "@/utils/basePath";
+import { useState } from "react";
+import { BomDetails } from "./BomDetails";
 
 export function ResourceTables({ resources, onDelete = null }) {
+  const [bomItem, setBomItem] = useState(null);
+
   return (
     <div className={styles.resouceTables}>
       {RESOURCE_ORDER.map((type) => {
@@ -24,21 +29,38 @@ export function ResourceTables({ resources, onDelete = null }) {
             <Table
               columns={RESOURCE_COLUMNS_MAP[type]}
               data={grouped}
-              rowButtons={
-                onDelete
-                  ? (row, index) => [
-                      {
-                        label: "Delete",
-                        className: "btn-delete",
-                        onClick: () => onDelete(type, index),
-                      },
-                    ]
-                  : undefined
-              }
+              rowButtons={(row, index = null) => {
+                const buttons = [];
+
+                if (row.resource_type === BASE_PATH.bom) {
+                  buttons.push({
+                    label: "View",
+                    onClick: () => {
+                      setBomItem(row);
+                    },
+                  });
+                }
+
+                if (onDelete) {
+                  buttons.push({
+                    label: "Delete",
+                    className: styles.btnDelete,
+                    onClick: () => onDelete(type, index),
+                  });
+                }
+
+                return buttons.length ? buttons : [];
+              }}
             />
           </div>
         );
       })}
+
+      {bomItem && (
+        <Modal title={`BOM Details`} onClose={() => setBomItem(null)}>
+          <BomDetails bomItem={bomItem} />
+        </Modal>
+      )}
     </div>
   );
 }
